Moved config to the database
Replaced the initialisation of the config file by two environment variables. MM_CONFIG make the database, the store location for the config file. MM_SERVICESETTINGS_LISTENADDRESS overrides the port stored in the database with the one that Heroku provides. This change allows for persistent configurations. Finally, since the location of config is an environment variable, the config flag is omitted.